Restructure directory tree: Flatten hotel-backend structure to a single layer
