--- hotel-backend/hotel-backend/src/test/java/com/example/hotel_backend/HotelBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.hotel_backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HotelBackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
