+++ hotel-backend/src/main/java/com/example/hotel_backend/HotelBackendApplication.java
@@ -1,0 +1,13 @@
+package com.example.hotel_backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HotelBackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HotelBackendApplication.class, args);
+	}
+
+}
